+++ src/FactoryGame.Domain/Content/MachinePortCatalog.cs
@@ -44,8 +44,6 @@ public static class MachinePortCatalog
                 new("out4", PortDirection.Out)
             ],
             ["SeaportConnector"] = [new("in", PortDirection.In), new("out", PortDirection.Out)],
-            ["SeaportIn"] = [new("out", PortDirection.Out)],
-            ["SeaportOut"] = [new("in", PortDirection.In)],
             ["Tank"] = [new("in", PortDirection.In), new("out", PortDirection.Out)],
             ["Junction"] = [new("in", PortDirection.In), new("out1", PortDirection.Out), new("out2", PortDirection.Out)],
             ["RateLimiter"] = [new("in", PortDirection.In), new("out", PortDirection.Out)]